+++ de.bund.bfr.knime.fsklab.nodes/js-src/de/bund/bfr/knime/fsklab/nodes/fskeditorjs/fskeditorjs.js
@@ -260,6 +260,28 @@ fskeditorjs = function() {
 						"input",inputHandler
 						);
 		/* execute a function presses a key on the keyboard: */
+		inp.addEventListener("focus", function(e) {
+			console.log("e");
+			var x = document.getElementById(this.id + "autocomplete-list");
+			if (x)
+				x = x.getElementsByTagName("div");
+			
+			if(x == null){
+				inputHandler(e);
+			}
+		});
+		inp.addEventListener("click", function(e) {
+			event.preventDefault(); // Let's stop this event.
+			event.stopPropagation(); // Really this time.
+			var x = document.getElementById(this.id + "autocomplete-list");
+			if (x)
+				x = x.getElementsByTagName("div");
+			
+			if(x == null){
+				inputHandler(e);
+			}
+			
+		});
 		inp.addEventListener("keydown", function(e) {
 			
 			var x = document.getElementById(this.id + "autocomplete-list");
@@ -937,12 +959,27 @@ fskeditorjs = function() {
 		});
 
 		$.each($("input[type='text']"), function(key, value) {
-
+			
 			$(value).removeAttr('class');
 			$(value).addClass('form-control');
 			$(value).parent().parent().removeAttr('class');
 			$(value).parent().parent().addClass('form-group');
 			fixInputCSS($(value));
+			$(value).focus(function(){
+			    if(this.value.length > 0){
+			    	if($(value).width() <  ((this.value.length + 1) * 12)){
+			    		this.style.width = ((this.value.length + 1) * 12) + 'px';
+			    		
+			    	}
+			    }
+
+			});
+			$(value).blur(function(){
+			    if(this.value.length > 0){
+			        this.style.width = "";
+			    }
+
+			});
 		});
 
 		$('.MuiFormLabel-root-100').css('font-size', '1.5rem');
@@ -999,8 +1036,10 @@ fskeditorjs = function() {
 				'patternConsumption', 'populationAge' ];
 		$("[aria-describedby*='tooltip-add']").click(function(event) {
 			currentArea = window.makeId($(this).attr('aria-label'));
-			//console.log(currentArea);
-
+			window.generalInformation = window.store1.getState().jsonforms.core.data;
+			window.scope = window.store2.getState().jsonforms.core.data;
+			window.modelMath =  window.store17.getState().jsonforms.core.data;
+			window.dataBackground =  window.store6.getState().jsonforms.core.data;
 			if ($.inArray(currentArea, StringObjectPopupsName) < 0) {
 				event.preventDefault(); // Let's stop this event.
 				event.stopPropagation(); // Really this time.
@@ -1020,6 +1059,9 @@ fskeditorjs = function() {
 				window.scrollTo(0, 0);
 			}
 		});
+
+
+		
 
 		autoCompleteCB = [ 'country', 'language', 'source', 'rights', 'format',
 				'software', 'languageWrittenIn', 'modelClass', 'basicProcess',
@@ -1221,7 +1263,13 @@ fskeditorjs = function() {
 			
 		})
 		//console.log(navigationMap);
+		$(".table-responsive").parent().css("flex-direction","unset");
+		$(".MuiGrid-typeContainer-1").css("display","inherit");
+		$(".MuiGrid-spacing-xs-16-22").css("display","flex");
+		
+		
 	}
+	
 	if (parent !== undefined && parent.KnimePageLoader !== undefined) {
 		parent.KnimePageLoader.autoResize(window, frameElement.id)
 	}